+++ scripts/embedding/xvm/xvm_logtalk_pl.ps1
@@ -4,11 +4,11 @@
 ##   This script creates a XVM logtalk.pl file with the Logtalk compiler and
 ##   runtime and optionally an application.pl file with a Logtalk application
 ## 
-##   Last updated on September 6, 2023
+##   Last updated on October 31, 2024
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   SPDX-FileCopyrightText: 2022 Hans N. Beck
-##   SPDX-FileCopyrightText: 2022 Paulo Moura <pmoura@logtalk.org>
+##   SPDX-FileCopyrightText: 2022-2024 Paulo Moura <pmoura@logtalk.org>
 ##   SPDX-License-Identifier: Apache-2.0
 ##   
 ##   Licensed under the Apache License, Version 2.0 (the "License");
@@ -39,9 +39,9 @@ param(
 	[String]$p = ($env:LOGTALKHOME + '\paths\paths.pl'),
 	[String]$s = ($env:LOGTALKHOME + '\scripts\embedding\settings-embedding-sample.lgt'),
 	[String]$l,
+	[String]$g = "true",
 	[Switch]$f,
 	[Switch]$x,
-	[String]$g = "true",
 	[Switch]$v,
 	[Switch]$h
 )
@@ -49,7 +49,7 @@ param(
 function Write-Script-Version {
 	$myFullName = $MyInvocation.ScriptName
 	$myName = Split-Path -Path $myFullName -leaf -Resolve
-	Write-Output ($myName + " 0.10")
+	Write-Output ($myName + " 0.11")
 }
 
 function Get-Logtalkhome {
@@ -94,10 +94,11 @@ function Write-Usage-Help() {
 	Write-Output "This script creates a XVM logtalk.pl file with the Logtalk compiler/runtime"
 	Write-Output "and an optional application.pl file from an application source code given"
 	Write-Output "its loader file. When embedding an application, this script also creates a"
-	Write-Output "loader.pl file for loading all generated Prolog and foreign library files."
+	Write-Output "loader.pl file for loading all generated Prolog and foreign library files,"
+	Write-Output "optionally calling a startup goal."
 	Write-Output ""
 	Write-Output "Usage:"
-	Write-Output ("  " + $myName + " [-c] [-d directory] [-t tmpdir] [-n name] [-p paths] [-s settings] [-l loader] [-f] [-x]")
+	Write-Output ("  " + $myName + " [-c] [-d directory] [-t tmpdir] [-n name] [-p paths] [-s settings] [-l loader] [-g goal] [-f] [-x]")
 	Write-Output ("  " + $myName + " -v")
 	Write-Output ("  " + $myName + " -h")
 	Write-Output ""
@@ -109,6 +110,7 @@ function Write-Usage-Help() {
 	Write-Output ("  -p library paths file (absolute path; default is " + $p + ")")
 	Write-Output ("  -s settings file (absolute path or 'none'; default is " + $s + ")")
 	Write-Output "  -l loader file for the application (absolute path)"
+	Write-Output "  -g startup goal for the application in canonical syntax (default is $g)"
 	Write-Output "  -f copy foreign library files loaded by the application"
 	Write-Output "  -x encrypt the generated logtalk.pl and application.pl files"
 	Write-Output ("  -v print version of " +  $myName)
@@ -281,8 +283,6 @@ if ($l -ne "") {
 		Get-Content |
 		Set-Content $d/application.pl
 
-	Set-Content -Path $d/loader.pl -Value ":- initialization(("
-
 	if ($x -eq $true) {
 		$GoalEncryptLogtalk = "encrypt_program('" + $d.Replace('\', '/') + "/logtalk.pl'),halt."
 		$GoalEncryptApplication = "encrypt_program('" + $d.Replace('\', '/') + "/application.pl'),halt."
@@ -298,8 +298,14 @@ if ($l -ne "") {
 		xvmpl --goal $LoaderParam
 	}
 
+	Set-Content -Path $d/loader.pl -Value ":- initialization(("
 	Add-Content -Path $d/loader.pl -Value  "	consult(logtalk),"
-	Add-Content -Path $d/loader.pl -Value  "	consult(application)"
+	if ($g -eq $true) {
+		Add-Content -Path $d/loader.pl -Value  "	consult(application)"
+	} else {
+		Add-Content -Path $d/loader.pl -Value  "	consult(application),"
+		Add-Content -Path $d/loader.pl -Value  "	$g"
+	}
 	Add-Content -Path $d/loader.pl -Value  "))."
 
 	Pop-Location