+++ RELEASE_NOTES.md
@@ -67,6 +67,15 @@ predicates.
 * ADDED: Additional tests for the de facto Prolog standard `format/2-3`
 predicates (`~p` control sequence).
 
+Installers and installation scripts
+-----------------------------------
+
+* UPDATED: The `trealla_logtalk_pl` embedding scripts to create a `loader.pl`
+file when the application loader option is provided.
+
+* UPDATED: The `trealla_logtalk_pl` and `xvm_logtalk_pl` embedding scripts
+to accept an optional startup application goal argument.
+
 
 3.84.0 - October 22, 2024
 =========================